Restructure Donor profile folders and files
