+++ src/routes/Donor-Profile/History Campaign/history.js
@@ -1,6 +1,6 @@
-import AvatarContainer from '../Left Side/Avatar'
-import Points from '../Left Side/Points'
-import Award from '../Left Side/Award'
+import AvatarContainer from '../../../components/Avatar, Point, Reward/Avatar'
+import Points from '../../../components/Avatar, Point, Reward/Points'
+import Award from '../../../components/Avatar, Point, Reward/Award'
 import HistoryContainer from './historyContainer'
 import styles from '../donor.module.css'
 import { Col, Row } from 'antd'
