+++ src/routes/Donor-Profile/Q&A/QnADonor.js
@@ -1,6 +1,6 @@
-import AvatarContainer from '../Left Side/Avatar'
-import Points from '../Left Side/Points'
-import Award from '../Left Side/Award'
+import AvatarContainer from '../../../components/Avatar, Point, Reward/Avatar'
+import Points from '../../../components/Avatar, Point, Reward/Points'
+import Award from '../../../components/Avatar, Point, Reward/Award'
 import QnAContainer from './QnAContainer'
 import styles from '../donor.module.css'
 import { Row, Col } from 'antd'
