+++ src/routes/Donor-Profile/donor-profile.js
@@ -1,6 +1,6 @@
 import { Col, Row } from 'antd';
 import { Outlet, Navigate } from 'react-router-dom';
-import { Navbar } from '../components/NavBar/navbar';
+import { Navbar } from '../../components/NavBar/navbar';
 // import { PageFooterBottom } from '../components/Footer/PageFooterBottom';
 function DonorProfile() {
     const user = JSON.parse(sessionStorage.getItem('user'))
